Tasks: - removed unused and duplicate  functions
